Fixed bugs, added cMenu
Fixed issue with moveableWindow not closing correctly causing error spam in the console.
attempted to speed up the updates of dragging a moveableWindow
Fixed styles of calendar objects to use uuids so duplicate widgets are not accidently configured simultaneously.
Added cMenu basics to allow for quick configuration of Menus using a context standard to generate the required elements.